+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessNoInitVisitor.java
@@ -299,10 +299,8 @@ public class NullnessNoInitVisitor extends BaseTypeVisitor<NullnessNoInitAnnotat
     /** Case 1: Check for null dereferencing. */
     @Override
     public Void visitMemberSelect(MemberSelectTree tree, Void p) {
-        if (ignoreCheckDeadCode) {
-            if (atypeFactory.isUnreachable(tree)) {
-                return super.visitMemberSelect(tree, p);
-            }
+        if (ignoreDeadCode && atypeFactory.isUnreachable(tree)) {
+            return super.visitMemberSelect(tree, p);
         }
         Element e = TreeUtils.elementFromUse(tree);
         if (e.getKind() == ElementKind.CLASS) {