+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -153,7 +153,7 @@ import javax.tools.Diagnostic;
 
     // Make checker ignore the expression in dead branch
     // org.checkerframework.framework.common.basetype.BaseTypeVisitor.shouldSkipUses
-    "ignoreCheckDeadCode",
+    "ignoreDeadCode",
 
     ///
     /// More sound (strict checking): enable errors that are disabled by default